Adding interface for pricer and moving to isolated namespace
