+++ src/main/java/com/xgen/interview/pricers/IPricer.java
@@ -1,0 +1,5 @@
+package com.xgen.interview.pricers;
+
+public interface IPricer {
+    public Integer getPrice(String itemType);
+}
