+++ src/main/java/com/xgen/interview/pricers/Pricer.java
@@ -1,4 +1,4 @@
-package com.xgen.interview;
+package com.xgen.interview.pricers;
 
 import java.util.HashMap;
 
@@ -6,7 +6,7 @@ import java.util.HashMap;
 /**
  * A stub implementation - for this exercise, you may disregard that this is incomplete.
  */
-public class Pricer {
+public class Pricer implements IPricer {
     HashMap<String, Integer> pricingDatabase = new HashMap<>(); // stub
 
     public Pricer() {
