+++ src/test/java/com/xgen/interview/ShoppingCartTest.java
@@ -1,7 +1,7 @@
 package com.xgen.interview;
 
 import com.xgen.interview.printers.StandardOutPrinter;
-import com.xgen.interview.Pricer;
+import com.xgen.interview.pricers.Pricer;
 import com.xgen.interview.ShoppingCart;
 
 import org.junit.Test;
